--- .markdownlint.json
@@ -1,9 +1,0 @@
-{
-  "MD033": false,
-  "MD041": false,
-  "MD045": false,
-  "MD003": false,
-  "MD013": {
-    "code_blocks": false
-  }
-}
