+++ pragma-node/src/handlers/mod.rs
@@ -26,7 +26,7 @@ use utoipa::{IntoParams, ToSchema};
 
 use pragma_common::types::{AggregationMode, DataType, Interval};
 
-use crate::types::timestamp::UnixTimestamp;
+use pragma_types::timestamp::UnixTimestamp;
 
 #[derive(Default, Debug, Deserialize, ToSchema, Clone, Copy)]
 pub enum EntryType {
