+++ pragma-node/src/handlers/onchain/get_history.rs
@@ -8,7 +8,7 @@ use utoipa::{IntoParams, ToResponse, ToSchema};
 use crate::infra::repositories::onchain_repository::history::{
     get_historical_entries_and_decimals, retry_with_routing, HistoricalEntryRaw,
 };
-use crate::types::timestamp::TimestampRange;
+use pragma_types::timestamp::TimestampRange;
 use crate::utils::{big_decimal_price_to_hex, PathExtractor};
 use crate::AppState;
