+++ pragma-node/src/handlers/publish_entry_ws.rs
@@ -5,9 +5,9 @@ use std::time::{Duration, SystemTime};
 use utoipa::ToSchema;
 
 use crate::handlers::create_entry::CreateEntryResponse;
-use crate::types::auth::{build_login_message, LoginMessage};
-use crate::types::entries::Entry;
-use crate::types::ws::{ChannelHandler, Subscriber, WebSocketError};
+use pragma_types::auth::{build_login_message, LoginMessage};
+use pragma_types::entries::Entry;
+use crate::utils::{ChannelHandler, Subscriber, WebSocketError};
 use crate::utils::{
     assert_login_is_valid, convert_entry_to_db, publish_to_kafka, validate_publisher,
 };
