+++ pragma-node/src/handlers/subscribe_to_entry.rs
@@ -15,9 +15,9 @@ use utoipa::{ToResponse, ToSchema};
 
 use crate::constants::starkex_ws::PRAGMA_ORACLE_NAME_FOR_STARKEX;
 use crate::infra::repositories::entry_repository::MedianEntryWithComponents;
-use crate::types::pricer::{IndexPricer, MarkPricer, Pricer};
-use crate::types::timestamp::UnixTimestamp;
-use crate::types::ws::{ChannelHandler, Subscriber, SubscriptionType};
+use crate::utils::pricer::{IndexPricer, MarkPricer, Pricer};
+use pragma_types::timestamp::UnixTimestamp;
+use crate::utils::{ChannelHandler, Subscriber, SubscriptionType};
 use crate::utils::{only_existing_pairs, sign_data, StarkexPrice};
 use crate::AppState;
