+++ pragma-node/src/utils/conversion.rs
@@ -3,8 +3,6 @@ use std::str::FromStr;
 use bigdecimal::BigDecimal;
 
 use pragma_entities::InfraError;
-use serde::{Deserialize, Deserializer};
-use starknet_crypto::Felt;
 
 pub fn convert_via_quote(
     a_price: BigDecimal,
@@ -48,33 +46,6 @@ pub fn format_bigdecimal_price(price: BigDecimal, decimals: u32) -> String {
         }
     }
     formatted_price
-}
-
-pub fn felt_from_decimal<'de, D>(deserializer: D) -> Result<Vec<Felt>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let s: Vec<String> = Vec::deserialize(deserializer)?;
-    Ok(s.iter().map(|s| Felt::from_dec_str(s).unwrap()).collect())
-}
-
-pub fn flexible_u128<'de, D>(deserializer: D) -> Result<u128, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    use serde::de::Error;
-
-    // Try deserializing to Value first to handle both formats
-    let value = serde_json::Value::deserialize(deserializer)?;
-
-    match value {
-        serde_json::Value::String(s) => u128::from_str(&s).map_err(D::Error::custom),
-        serde_json::Value::Number(n) => {
-            let s = n.to_string();
-            u128::from_str(&s).map_err(D::Error::custom)
-        }
-        _ => Err(D::Error::custom("expected string or number")),
-    }
 }
 
 /// Converts two currencies pairs to a new routed pair id.
