+++ pragma-types/README.md
@@ -1,0 +1,3 @@
+# Pragma Types
+
+Common types available for external use when interacting with Pragma in rust.
