+++ pragma-types/src/hex_hash.rs
