+++ pragma-types/src/lib.rs
@@ -1,0 +1,10 @@
+// Re-export all modules
+pub mod auth;
+pub mod entries;
+pub mod hex_hash;
+pub mod timestamp;
+pub mod typed_data;
+pub mod utils;
+
+// Re-export commonly used types for convenience
+pub use entries::Entry;
